+++ rules/index.ts
@@ -45,10 +45,8 @@ const rules: Rule<ITheme>[] = [
     'border-style': 'solid',
     'border-width': '1px',
   }],
-  // ['round-full', { 'border-radius': '999999px' }],
-  // size.rule('round', 'border-radius'),
-  ...size.cornerRules('round', '', 'border', 'radius'),
-  ...logical.cornerRules('round', 'full', 'border', 'radius', (pref, prop) => simpleRule(pref, prop, '999999px')),
+  ...size.cornerRules('rounded', '', 'border', 'radius'),
+  ...logical.cornerRules('rounded', 'full', 'border', 'radius', (pref, prop) => simpleRule(pref, prop, '999999px')),
   ...logical.edgeRules('border', 'color', 'border', 'color', colorRule),
   ...logical.edgeRules('border', 'color-fg', 'border', 'color', fgColorRule),